Extract LoadInsertable, rename LoadDocumentWorkflow to LoadGroupWorkflow
Splits per-element loading out of the group workflow into its own class,
LoadInsertable (load-insertable.ts): it owns everything about loading and
persisting one insertable independently of the rest of the group -
configuration, a discrete vendor-parsing step, thumbnails, insert-and-fasten,
and finally its own D1 write. Each insertable now writes its own row rather
than all elements landing in one batch at the end, so one element's
exhausted retries no longer block the others or the group-level finalize
(group row upsert, stale-insertable cleanup, search rebuild) - a stale
microversionId is simply left in place, which the existing
getInsertablesToReload check already treats as needing another reload.

LoadDocumentWorkflow is renamed LoadGroupWorkflow (load-group-workflow/
load-group.ts), trimmed down to group-level orchestration now that the
per-element logic has moved out. Its version-fetch step also simplifies to
a single getDocument call, since the DB no longer needs the version's
name/createdAt (see below) - just recentVersion.id.

Also renames the groups/insertables schema's instanceId column to
versionId, and drops insertables.versionName/versionCreatedAt (unused
outside the backend, never displayed). conflictUpdateSet and the
thumbnail-upload retry helper move to shared homes (db.ts,
routes/thumbnails.ts) now that both the group and insertable loads need
them.
