+++ drizzle/0003_rename_version_id_drop_version_fields.sql
@@ -1,0 +1,4 @@
+ALTER TABLE `groups` RENAME COLUMN `instance_id` TO `version_id`;--> statement-breakpoint
+ALTER TABLE `insertables` RENAME COLUMN `instance_id` TO `version_id`;--> statement-breakpoint
+ALTER TABLE `insertables` DROP COLUMN `version_name`;--> statement-breakpoint
+ALTER TABLE `insertables` DROP COLUMN `version_created_at`;
